(#65) Reorganizing models.py as folder
