+++ d2qc/d2qc/data/models/data_set.py
@@ -1,308 +1,12 @@
-from django.contrib.gis.db import models
-from django.forms import ModelForm
-from django.contrib.auth.models import User
-from django.conf import settings
-from django.core.cache import cache
-from django.contrib.gis.geos import Point
-import os.path
-from django.utils import timezone
-import logging
-import re
-from decimal import Decimal
 import math
-from django.db import connection
 import gsw
+from django.contrib.gis.db import models
+from django.contrib.auth.models import User
+from django.db import connection
+from django.core.cache import cache
 import glodap.util.interp as interp
 import glodap.util.stats as stats
 import glodap.util.geo as geo
-import glodap.util.excread as excread
-import pandas as pd
-import statistics as stat
-
-class Profile(models.Model):
-    class Meta:
-        db_table = 'd2qc_user_profile'
-    user = models.OneToOneField(
-        User,
-        related_name='profile',
-        on_delete = models.CASCADE,
-        blank=True,
-        null=True,
-        editable=False
-    )
-    min_depth = models.IntegerField(default=1500)
-    crossover_radius = models.IntegerField(default=200000)
-    created = models.DateTimeField(auto_now_add=True)
-    updated = models.DateTimeField(auto_now=True)
-
-class DataFile(models.Model):
-    class Meta:
-        db_table = 'd2qc_data_files'
-
-    # Make sure files get unique filenames
-    def file_store_path(self, filename):
-        # clear filename of illegal characters
-        filename = re.sub('[^a-zA-Z0-9\.\-\_]', '', filename)
-        id = self.owner.id
-        path = os.path.join(
-                settings.DATA_FOLDER,
-                'UID_{}'.format(id)
-        )
-        i = 0
-        name = '{}__{}'.format(i, filename)
-        while os.path.isfile(os.path.join(path, name)):
-            i += 1
-            name = '{}__{}'.format(i, filename)
-        return os.path.join(path, name)
-
-    id = models.AutoField(primary_key=True)
-    filepath = models.FileField(
-            upload_to=file_store_path,
-            blank=True,
-            null=True
-    )
-    name = models.CharField(max_length=255, blank=True)
-    description = models.CharField(max_length=255, blank=True)
-    headers = models.TextField(blank=True)
-    created = models.DateTimeField(auto_now_add=True)
-    updated = models.DateTimeField(auto_now=True)
-    owner = models.ForeignKey(
-            User,
-            on_delete = models.PROTECT,
-            blank=True,
-            null=True,
-            editable=False
-    )
-    import_errors = models.TextField(blank=True)
-    import_started = models.DateTimeField(null=True)
-    import_finnished = models.DateTimeField(null=True)
-
-    # Messages reported while importing data
-    _messages = []
-
-    def __str__(self):
-        return self.name if self.name else self.filepath
-
-    # Delete files as object is deleted
-    def delete(self):
-        # Dont delete file if file has related data set(s)
-        if not DataSet.objects.filter(data_file_id=self.id).exists():
-            self.filepath.delete()
-        super().delete()
-
-    def read_file(self):
-        filearray = []
-        if self.filepath:
-            path = os.path.join(
-                settings.BASE_DIR,
-                str(self.filepath),
-            )
-            try:
-                with open(path, encoding="utf-8") as excfile:
-                    filearray = excfile.readlines()
-            except:
-                with open(path, encoding="iso-8859-1") as excfile:
-                    filearray = excfile.readlines()
-        return filearray
-
-    def _write_messages(self, append = False, save = False):
-        if append:
-            self.import_errors += '\n'.join(self._messages)
-        else:
-            self.import_errors = '\n'.join(self._messages)
-        if save:
-            self.save()
-
-    def import_data(self):
-        """
-        Import data from this data file. Abort if data existsself. Import errors
-        are appended to the import_errors field.
-
-        return True if data was imported, else return False
-        """
-        if self.import_started:
-            self._messages = ["File data already beeing imported"]
-            self._write_messages(append = True, save = True)
-            return False
-
-        self.import_started = timezone.now()
-        self.save()
-
-        datagrid = excread.excread(str(self.filepath))
-
-        MANDATORY_VARS = (
-            'EXPOCODE', 'EXC_DATETIME', 'EXC_CTDDEPTH', 'STNNBR', 'LATITUDE',
-            'LONGITUDE',
-        )
-
-        # Variables not to be treated as data variables
-        IGNORE = (
-            'EXPOCODE', 'EXC_DATETIME', 'EXC_CTDDEPTH', 'STNNBR', 'SECT_ID', 'DATE',
-            'TIME', 'LATITUDE', 'LONGITUDE', 'BTLNBR', 'BTLNBR_FLAG_W',
-            'SAMPNO', 'CASTNO', 'CTDDEPTH', 'CTDDEP', 'HOUR', 'MINUTE', 'DEPTH'
-        )
-
-        QC_SUFFIX = '_FLAG_W'
-
-
-        # Check all mandatory variables are there
-        depth = ''
-        stnnbr = ''
-        castno = ''
-        data_set = None
-        station = None
-        cast = None
-        depth = None
-        # Raise an exception if mandatory columns are missing
-        if not all(key in datagrid.columns for key in MANDATORY_VARS):
-            message = "Data file missing some mandatory column: {}".format(
-                ', '.join(MANDATORY_VARS)
-            )
-            self._messages.append(message)
-            self._write_messages()
-            self.import_finnished = timezone.now()
-            self.save()
-            return False
-
-        # Import data types
-        missing_vars = []
-        data_types = {str(type_):type_ for type_ in DataType.objects.all()}
-        for var in datagrid.columns:
-            if var in IGNORE:
-                continue
-            if var.endswith(QC_SUFFIX):
-                continue
-            if var not in data_types:
-                missing_vars.append(var)
-
-        if missing_vars:
-            message = """There where variables in the dataset that are not defined in
-            the system. These cannot be handled. An administrator has to add
-            the variables as data types for them to be treated. Unhandled
-            variables in the data set: {}
-            """.format(
-                ', '.join(missing_vars)
-            )
-            self._messages.append(message)
-
-        missing_depth_warning = False # Indicate missing depth already warned
-        missing_position_warning = False
-        for i, expo in enumerate(datagrid['EXPOCODE']):
-            if not data_set or expo != data_set.expocode:
-                # Add new dataset
-                data_set = DataSet(
-                    expocode=expo,
-                    is_reference = False,
-                    data_file = self,
-                    owner = self.owner
-                )
-                if DataSet.objects.filter(
-                            expocode=expo,
-                            owner=self.owner
-                ).exists():
-                    # TODO Support files with multiple datasets, where one or
-                    # more might already exist in database, but not all.
-                    message = 'Dataset {} already exists for this user'. format(
-                        expo
-                    )
-                    self._messages = [message]
-                    self._write_messages()
-                    self.import_finnished = timezone.now()
-                    self.save()
-                    return False
-
-                data_set.save()
-                station = None
-                cast = None
-                depth = None
-            if not station or datagrid['STNNBR'][i] != station.station_number:
-                longitude = datagrid['LONGITUDE'][i]
-                latitude = datagrid['LATITUDE'][i]
-                if math.isnan(longitude) or math.isnan(latitude):
-                    if missing_position_warning:
-                        continue
-                    # Warning and dont insert if depth is NaN
-                    message = """Latitude or longitude is nan on line {}.
-                    Station will not be added when position is missing.
-                    Subsequent missing position errors are supressed for this
-                    file.
-                    """.format(i)
-                    self._messages.append(message)
-                    missing_position_warning = True
-                    continue
-                # Add new station
-                station = Station(
-                        data_set = data_set,
-                        position = Point(longitude, latitude),
-                        station_number = datagrid['STNNBR'][i]
-                )
-                station.save()
-                cast = None
-                depth = None
-            if (
-                    not cast or
-                    ('CASTNO' in datagrid and datagrid['CASTNO'][i] != cast.cast)
-            ):
-                # Add new cast
-                cast_ = 1
-                if 'CASTNO' in datagrid:
-                    cast_ = datagrid['CASTNO'][i]
-                cast = Cast(
-                        station = station,
-                        cast = cast_
-                )
-                cast.save()
-                depth = None
-
-            if (
-                    not depth
-                    or depth.depth != datagrid['EXC_CTDDEPTH'][i]
-                    or (
-                        'BTLNBR' in datagrid
-                        and depth.bottle != datagrid['BTLNBR'][i]
-                    )
-            ):
-                if math.isnan(datagrid['EXC_CTDDEPTH'][i]):
-                    if missing_depth_warning:
-                        continue
-                    # Warning and dont insert if depth is NaN
-                    message = """Depth is nan on line {}. Data will not be added when
-                            depth is nan. Subsequent missing depth errors are
-                            supressed for this file.
-                            """.format(i)
-                    self._messages.append(message)
-                    missing_depth_warning = True
-                    continue
-
-                # Add new depth
-                btlnbr = datagrid.get('BTLNBR', False)
-                depth = Depth(
-                        cast = cast,
-                        depth = datagrid['EXC_CTDDEPTH'][i],
-                        bottle = 1 if btlnbr is False else btlnbr[i],
-                        date_and_time = datagrid['EXC_DATETIME'][i],
-                )
-                depth.save()
-            for key in datagrid.columns:
-                if key in IGNORE:
-                    continue
-                if not key in data_types:
-                    # Variable not found in database
-                    continue
-                qc_flag = None
-                if key + QC_SUFFIX in datagrid:
-                    qc_flag = int(datagrid[key + QC_SUFFIX][i])
-                value = DataValue(
-                        depth = depth,
-                        value = datagrid[key][i],
-                        qc_flag = qc_flag,
-                        data_type = data_types[key]
-                )
-                value.save()
-        self._write_messages()
-        self.import_finnished = timezone.now()
-        self.save()
-        return True
 
 class DataSet(models.Model):
     class Meta:
@@ -1109,131 +813,3 @@ class DataSet(models.Model):
             cache.set(cache_key, result)
 
         return result
-
-
-class DataType(models.Model):
-    class Meta:
-        db_table = 'd2qc_data_types'
-        unique_together = ('original_label', 'identifier', 'data_unit')
-
-    id = models.AutoField(primary_key=True)
-    data_unit = models.ForeignKey(
-        'DataUnit',
-        on_delete = models.CASCADE,
-        blank = True,
-        null = True,
-    )
-    identifier = models.CharField(max_length=20, default='', blank=True)
-    prefLabel = models.CharField(max_length=255, default='', blank=True)
-    altLabel = models.CharField(max_length=255, default='', blank=True)
-    definition = models.CharField(max_length=255, default='', blank=True)
-    original_label = models.CharField(max_length=20, default='')
-    offset_type = models.ForeignKey(
-        'OffsetType',
-        on_delete = models.PROTECT,
-        blank = True,
-        null = True,
-        default = 1,
-    )
-    created = models.DateTimeField(auto_now_add=True)
-    updated = models.DateTimeField(auto_now=True)
-    def __str__(self):
-        label = self.original_label if self.original_label else self.identifier
-        return label
-
-class OffsetType(models.Model):
-    class Meta:
-        db_table = 'd2qc_offset_types'
-    id = models.AutoField(primary_key=True)
-    name = models.CharField(max_length=20, default='', blank=True)
-    description = models.CharField(max_length=255, default='', blank=True)
-    def __str__(self):
-        return self.name
-
-class DataUnit(models.Model):
-    class Meta:
-        db_table = 'd2qc_data_units'
-
-    id = models.AutoField(primary_key=True)
-    identifier = models.CharField(max_length=20, default='', blank=True)
-    prefLabel = models.CharField(max_length=255, default='', blank=True)
-    altLabel = models.CharField(max_length=255, default='', blank=True)
-    definition = models.CharField(max_length=255, default='', blank=True)
-    original_label = models.CharField(max_length=20, default='')
-    created = models.DateTimeField(auto_now_add=True)
-    updated = models.DateTimeField(auto_now=True)
-    def __str__(self):
-        label = self.original_label if self.original_label else self.identifier
-        return label
-
-class Station(models.Model):
-    class Meta:
-        db_table = 'd2qc_stations'
-        unique_together = (
-                'position',
-                'station_number',
-                'data_set',
-        )
-        ordering = ['data_set', 'station_number']
-
-    id = models.AutoField(primary_key=True)
-    data_set = models.ForeignKey('DataSet', related_name='stations',
-            on_delete=models.CASCADE)
-    position = models.PointField(srid=4326, spatial_index=True)
-    station_number = models.IntegerField(null=True)
-    created = models.DateTimeField(auto_now_add=True)
-    updated = models.DateTimeField(auto_now=True)
-    def __str__(self):
-        label = str(self.station_number)
-        return label
-
-class Cast(models.Model):
-    class Meta:
-        db_table = 'd2qc_casts'
-
-    id = models.AutoField(primary_key=True)
-    station = models.ForeignKey('Station',
-            related_name='casts', on_delete=models.CASCADE)
-    cast = models.IntegerField(null=True)
-    created = models.DateTimeField(auto_now_add=True)
-    updated = models.DateTimeField(auto_now=True)
-
-class Depth(models.Model):
-    class Meta:
-        db_table = 'd2qc_depths'
-        indexes = [
-            models.Index(fields=['depth']),
-        ]
-        unique_together = ( 'cast', 'depth', 'bottle', 'date_and_time')
-        ordering = ['depth']
-
-    id = models.AutoField(primary_key=True)
-    cast = models.ForeignKey('Cast',
-            related_name='depths', on_delete=models.CASCADE)
-    depth = models.DecimalField(max_digits=8, decimal_places=3)
-    date_and_time = models.DateTimeField()
-    bottle = models.IntegerField(default=0)
-    created = models.DateTimeField(auto_now_add=True)
-    updated = models.DateTimeField(auto_now=True)
-
-class DataValue(models.Model):
-    class Meta:
-        db_table = 'd2qc_data_values'
-        unique_together = ('depth', 'data_type')
-
-    id = models.AutoField(primary_key=True)
-    depth = models.ForeignKey('Depth', related_name='data_values',
-            on_delete=models.CASCADE)
-    data_type = models.ForeignKey('DataType', on_delete=models.CASCADE)
-    value = models.DecimalField(max_digits=16, decimal_places=8)
-    qc_flag = models.IntegerField(blank=True, null=True)
-    qc2_flag = models.IntegerField(blank=True, null=True)
-    created = models.DateTimeField(auto_now_add=True)
-    updated = models.DateTimeField(auto_now=True)
-    def save(self, *args, **kwargs):
-        try:
-            v = Decimal(self.value)
-        except ValueError:
-            logger.error('Data Value ' + self.value + ' is not a float value')
-        if not math.isnan(v) and int(v) != -9999:
-            super(DataValue, self).save(*args, **kwargs)
