+++ d2qc/d2qc/data/models/data_type.py
@@ -1,0 +1,32 @@
+from django.contrib.gis.db import models
+
+
+class DataType(models.Model):
+    class Meta:
+        db_table = 'd2qc_data_types'
+        unique_together = ('original_label', 'identifier', 'data_unit')
+
+    id = models.AutoField(primary_key=True)
+    data_unit = models.ForeignKey(
+        'DataUnit',
+        on_delete = models.CASCADE,
+        blank = True,
+        null = True,
+    )
+    identifier = models.CharField(max_length=20, default='', blank=True)
+    prefLabel = models.CharField(max_length=255, default='', blank=True)
+    altLabel = models.CharField(max_length=255, default='', blank=True)
+    definition = models.CharField(max_length=255, default='', blank=True)
+    original_label = models.CharField(max_length=20, default='')
+    offset_type = models.ForeignKey(
+        'OffsetType',
+        on_delete = models.PROTECT,
+        blank = True,
+        null = True,
+        default = 1,
+    )
+    created = models.DateTimeField(auto_now_add=True)
+    updated = models.DateTimeField(auto_now=True)
+    def __str__(self):
+        label = self.original_label if self.original_label else self.identifier
+        return label
