+++ d2qc/d2qc/data/models/data_value.py
@@ -1,0 +1,23 @@
+from django.contrib.gis.db import models
+
+class DataValue(models.Model):
+    class Meta:
+        db_table = 'd2qc_data_values'
+        unique_together = ('depth', 'data_type')
+
+    id = models.AutoField(primary_key=True)
+    depth = models.ForeignKey('Depth', related_name='data_values',
+            on_delete=models.CASCADE)
+    data_type = models.ForeignKey('DataType', on_delete=models.CASCADE)
+    value = models.DecimalField(max_digits=16, decimal_places=8)
+    qc_flag = models.IntegerField(blank=True, null=True)
+    qc2_flag = models.IntegerField(blank=True, null=True)
+    created = models.DateTimeField(auto_now_add=True)
+    updated = models.DateTimeField(auto_now=True)
+    def save(self, *args, **kwargs):
+        try:
+            v = Decimal(self.value)
+        except ValueError:
+            logger.error('Data Value ' + self.value + ' is not a float value')
+        if not math.isnan(v) and int(v) != -9999:
+            super(DataValue, self).save(*args, **kwargs)
